requirements.txt
openai>=1.30.0
faiss-cpu
numpy
pydantic
tqdm
pyyaml
regex
python-dotenv
